Added sound effects for the menu navigation, robot launch, robot clean, robot power down (die), as well as background music and seagull ambience.
Fixed a few interface bugs with the subsystem menu, navigation should work properly for both up and down directions. Fixed an issue with some numbers not being able to be adjusted (energy level) properly.

Added a 'glow' effect to the subsystem selector to make it a bit more lively and not seem frozen looking.